Fixes
- Removing a blueprint was not calculating the buildCostMult into the refund value, causing lost credits
- Lights weren't initializing to on at game start since recent changes
- Added more easy Crabs to 2nd planets waves
- Made the bat wave 1.5 instead of two (also changed numberToSpawn to a float, for the subwaves)

+++ Mutant Mayhem/Assets/_Scripts/Waves/SubWaves/SubWaveSO.cs
@@ -7,5 +7,5 @@ using UnityEngine;
 public class SubWaveSO : ScriptableObject
 {
     public List<GameObject> enemyPrefabList;
-    public List<int> numberToSpawn;
+    public List<float> numberToSpawn;
 }
